Struct pg_ecs::prelude::query::ChunkView

pub struct ChunkView<'a, F>
where
    F: Fetch,
{
    archetype: &'a Archetype,
    fetch: F,
}

Provides access to slices of components for entities which have the same component layout.

A single index in any of the slices contained in a chunk belong to the same entity.

Fields
archetype: &'a Archetypefetch: F
Implementations

impl<'a, F> ChunkView<'a, F>
where F: Fetch,

pub fn archetype(&self) -> &Archetype

Returns the archetype that all entities in the chunk belong to.

pub fn component_slice<T>(&self) -> Option<&[T]>
where T: Component,

Returns a slice of components.

May return None if the chunk’s view does not declare access to the component type.

pub fn component_slice_mut<T>(&mut self) -> Option<&mut [T]>
where T: Component,

Returns a mutable slice of components.

May return None if the chunk’s view does not declare access to the component type.

pub fn into_components(self) -> <F as Fetch>::Data

Converts the chunk into a tuple of it’s inner slices.

Examples

let mut query = <(Entity, Read<A>, Write<B>, TryRead<C>, TryWrite<D>)>::query();
for chunk in query.iter_chunks_mut(&mut world) {
    let slices: (&[Entity], &[A], &mut [B], Option<&[C]>, Option<&mut [D]>) = chunk.into_components();       
}

pub fn get_components(&self) -> <F as Fetch>::Data
where F: ReadOnlyFetch,

Converts the chunk into a tuple of it’s inner slices.

Only usable with views who’s elements are all read-only.

Examples

let mut query = <(Entity, Read<A>, TryRead<B>)>::query();
for chunk in query.iter_chunks_mut(&mut world) {
    let slices: (&[Entity], &[A], Option<&[B]>) = chunk.get_components();       
}

pub fn into_iter_entities( self, ) -> impl Iterator<Item = (Entity, <F as IntoIndexableIter>::Item)> + 'a
where <F as IntoIndexableIter>::IntoIter: 'a,

Converts the chunk into an iterator which yields tuples of (Entity, components).