Trait pg_ecs::prelude::query::View

pub trait View<'data>: Sized + DefaultFilter {
    type Element: Send + Sync + 'data;
    type Fetch: Fetch<Item = Self::Element> + IntoIndexableIter + 'data;
    type Iter: Iterator<Item = Option<Self::Fetch>> + 'data;
    type Read: AsRef<[ComponentTypeId]>;
    type Write: AsRef<[ComponentTypeId]>;

    // Required methods
    unsafe fn fetch(
        components: &'data Components,
        archetypes: &'data [Archetype],
        query: QueryResult<'data>,
    ) -> Self::Iter;
    fn validate();
    fn validate_access(access: &ComponentAccess<'_>) -> bool;
    fn reads_types() -> Self::Read;
    fn writes_types() -> Self::Write;
    fn reads<T>() -> bool
       where T: Component;
    fn writes<T>() -> bool
       where T: Component;
    fn requires_permissions() -> Permissions<ComponentTypeId>;
}

A type which can pull entitiy data out of a world.

Required Associated Types

type Element: Send + Sync + 'data

The type of component references returned.

type Fetch: Fetch<Item = Self::Element> + IntoIndexableIter + 'data

The fetch type yielded for each archetype.

type Iter: Iterator<Item = Option<Self::Fetch>> + 'data

The iterator type which pulls entity data out of a world.

type Read: AsRef<[ComponentTypeId]>

Contains the type IDs read by the view.

type Write: AsRef<[ComponentTypeId]>

Contains the type IDs written by the view.

Required Methods

unsafe fn fetch( components: &'data Components, archetypes: &'data [Archetype], query: QueryResult<'data>, ) -> Self::Iter

Creates an iterator which will yield slices of entity data for each archetype.

Safety

This method may return mutable references to entity data via shared world references. The caller must ensure that no two view iterators are alive at the same time which access any components in a manor which may cause mutable aliasing.

fn validate()

Determines if this view type is valid. Panics if checks fail.

fn validate_access(access: &ComponentAccess<'_>) -> bool

Returns true if the given component access includes all permissions required by the view.

fn reads_types() -> Self::Read

Returns the component types read by the view.

fn writes_types() -> Self::Write

Returns the component types written to by the view.

fn reads<T>() -> bool
where T: Component,

Returns true if the view reads the specified data type.

fn writes<T>() -> bool
where T: Component,

Returns true if the view writes to the specified data type.

fn requires_permissions() -> Permissions<ComponentTypeId>

Returns a permissions struct declaring the component accesses required by the view.

Object Safety

This trait is not object safe.