Struct pg_ecs::prelude::Input

pub struct Input {
    keys_down: HashSet<VirtualKeyCode>,
    scancodes: HashSet<u32>,
    mouse_buttons: HashSet<usize>,
    mouse_pixel: (f64, f64),
    mouse_tile: Vec<(i32, i32)>,
    pub(crate) use_events: bool,
    event_queue: VecDeque<BEvent>,
    scale_factor: f64,
}

Represents the current input state. The old key/mouse fields remain available for compatibility.

Fields
keys_down: HashSet<VirtualKeyCode>scancodes: HashSet<u32>mouse_buttons: HashSet<usize>mouse_pixel: (f64, f64)mouse_tile: Vec<(i32, i32)>use_events: boolevent_queue: VecDeque<BEvent>scale_factor: f64
Implementations

impl Input

pub fn is_key_pressed(&self, key: VirtualKeyCode) -> bool

Checks to see if a key is pressed. True if it is, false otherwise.

pub fn is_scancode_pressed(&self, scan_code: u32) -> bool

Checks to see if a key is pressed by scancode. True if it is, false if it isnโt.

pub fn is_mouse_button_pressed(&self, button_num: usize) -> bool

Checks to see if a mouse button is pressed. 0 = left, 1 = right, 2 = middle, etc. is the convention for button numbering.

pub fn mouse_button_pressed_set(&self) -> &HashSet<usize>

HashSet of pressed mouse buttons

pub fn scan_code_pressed_set(&self) -> &HashSet<u32>

HashSet of pressed keyboard scan codes

pub fn key_pressed_set(&self) -> &HashSet<VirtualKeyCode>

HashSet of pressed keyboard scan codes

pub fn mouse_tile_pos(&self, console: usize) -> (i32, i32)

Returns the current mouse position (0,0 if there isnโt one yet), in TILE coordinates for the specified console layer.

pub fn mouse_tile(&self, console: usize) -> Point

Returns the current mouse position (0,0 if there isnโt one yet), in TILE coordinates for the specified console layer in Point format.

pub fn mouse_pixel_pos(&self) -> (f64, f64)

Return the current mouse position in pixels.

pub fn activate_event_queue(&mut self)

Call this to enable the event queue. Otherwise, events will not be tracked/stored outside of the HAL setup (to avoid continually filling a buffer that isnโt being used).

pub fn pop(&mut self) -> Option<BEvent>

Pop a single event from the event queue. Returns None if there arenโt any events.

pub fn for_each_message<F>(&mut self, action: F)
where F: FnMut(BEvent),

Provides a for_each function for all messages in the queue.

pub fn get_scale_factor(&self) -> f64

Gets the current scale factor