Struct pg_ecs::prelude::systems::Executor

pub struct Executor {
    systems: Vec<SystemBox>,
    static_dependants: Vec<Vec<usize>>,
    dynamic_dependants: Vec<Vec<usize>>,
    static_dependency_counts: Vec<AtomicUsize>,
    awaiting: Vec<AtomicUsize>,
}

Executes a sequence of systems, potentially in parallel, and then commits their command buffers.

Systems are provided in execution order. When the parallel feature is enabled, the Executor may run some systems in parallel. The order in which side-effects (e.g. writes to resources or entities) are observed is maintained.

Fields
systems: Vec<SystemBox>static_dependants: Vec<Vec<usize>>dynamic_dependants: Vec<Vec<usize>>static_dependency_counts: Vec<AtomicUsize>awaiting: Vec<AtomicUsize>
Implementations

impl Executor

pub fn new(systems: Vec<Box<dyn ParallelRunnable>>) -> Executor

Constructs a new executor for all systems to be run in a single stage.

Systems are provided in the order in which side-effects (e.g. writes to resources or entities) are to be observed.

pub fn into_vec(self) -> Vec<Box<dyn ParallelRunnable>>

Converts this executor into a vector of its component systems.

pub fn execute(&mut self, world: &mut World, resources: &mut Resources)

Executes all systems and then flushes their command buffers.

pub fn run_systems(&mut self, world: &mut World, resources: &UnsafeResources)

Executes all systems, potentially in parallel.

Ordering is retained in so far as the order of observed resource and component accesses is maintained.

Call from within rayon::ThreadPool::install() to execute within a specific thread pool.

pub fn flush_command_buffers(&mut self, world: &mut World)

Flushes the recorded command buffers for all systems.