Struct pg_ecs::prelude::systems::System

pub struct System<R, Q, F> {
    name: Option<SystemId>,
    _resources: ResourceMarker<R>,
    queries: Q,
    run_fn: F,
    archetypes: ArchetypeAccess,
    access: SystemAccess,
    command_buffer: HashMap<WorldId, CommandBuffer>,
}

The concrete type which contains the system closure provided by the user. This struct should not be constructed directly, and instead should be created using SystemBuilder.

Implements Schedulable which is consumable by the StageExecutor, executing the closure.

Also handles caching of archetype information, as well as maintaining the provided information about what queries this system will run and, as a result, its data access.

Fields
name: Option<SystemId>_resources: ResourceMarker<R>queries: Qrun_fn: Farchetypes: ArchetypeAccessaccess: SystemAccesscommand_buffer: HashMap<WorldId, CommandBuffer>