Trait pg_ecs::prelude::systems::QuerySet

pub trait QuerySet: Send + Sync {
    // Required method
    fn filter_archetypes(&mut self, world: &World, archetypes: &mut BitSet);
}

Provides an abstraction across tuples of queries for system closures.

Required Methods

fn filter_archetypes(&mut self, world: &World, archetypes: &mut BitSet)

Evaluates the queries and records which archetypes they require access to into a bitset.