Trait pg_ecs::prelude::systems::ResourceSet

pub trait ResourceSet<'a> {
    type Result: 'a;

    // Required method
    unsafe fn fetch_unchecked(resources: &'a UnsafeResources) -> Self::Result;

    // Provided methods
    fn fetch_mut(resources: &'a mut Resources) -> Self::Result { ... }
    fn fetch(resources: &'a Resources) -> Self::Result
       where Self: ReadOnly { ... }
}

Trait which is implemented for tuples of resources and singular resources. This abstracts fetching resources to allow for ergonomic fetching.

Example:

struct TypeA(usize);
struct TypeB(usize);

let mut resources = Resources::default();
resources.insert(TypeA(55));
resources.insert(TypeB(12));

{
    let (a, mut b) = <(Read<TypeA>, Write<TypeB>)>::fetch_mut(&mut resources);
    assert_ne!(a.0, b.0);
    b.0 = a.0;
}

{
    let (a, b) = <(Read<TypeA>, Read<TypeB>)>::fetch(&resources);
    assert_eq!(a.0, b.0);
}

Required Associated Types

type Result: 'a

The resource reference returned during a fetch.

Required Methods

unsafe fn fetch_unchecked(resources: &'a UnsafeResources) -> Self::Result

Fetches all defined resources, without checking mutability.

Safety

It is up to the end user to validate proper mutability rules across the resources being accessed.

Provided Methods

fn fetch_mut(resources: &'a mut Resources) -> Self::Result

Fetches all defined resources.

fn fetch(resources: &'a Resources) -> Self::Result
where Self: ReadOnly,

Fetches all defined resources.

Object Safety

This trait is not object safe.