Trait pg_ecs::prelude::Fetch

pub trait Fetch: IntoIndexableIter + Send + Sync {
    type Data;

    // Required methods
    fn into_components(self) -> Self::Data;
    fn find<T>(&self) -> Option<&[T]>
       where T: 'static;
    fn find_mut<T>(&mut self) -> Option<&mut [T]>
       where T: 'static;
    fn version<T>(&self) -> Option<u64>
       where T: Component;
    fn accepted(&mut self);
}

A type which holds onto a slice of entitiy data retrieved from a single archetype.

Required Associated Types

type Data

The inner data representation fetched from the archetype. Typically a slice reference.

Required Methods

fn into_components(self) -> Self::Data

Converts the fetch into the retrieved component slices

fn find<T>(&self) -> Option<&[T]>
where T: 'static,

Tries to find a slice of components, if this fetch contains the requested component type.

fn find_mut<T>(&mut self) -> Option<&mut [T]>
where T: 'static,

Tries to find a mutable slice of components, if this fetch contains the requested component type.

fn version<T>(&self) -> Option<u64>
where T: Component,

Tries to find the component slice version of a component, if this fetch contains the requested component type.

fn accepted(&mut self)

Indicates that the archetype is going to be provided to the user. Component slice versions are incremented here.

Object Safety

This trait is not object safe.